DocBotApp/DocBotApp/backend/requirements.txt
fastapi
uvicorn
python-multipart
passlib[bcrypt]
PyJWT
python-docx
